in delete endpoint add validation and fix wrong status code in put

+++ backend/controllers/product.controller.js
@@ -58,7 +58,7 @@ export const putProduct = async (req, res) => {
 
     try {
         const updatedProduct = await Product.findByIdAndUpdate(id, product, { new: true });
-        res.status(500).json({
+        res.status(200).json({
             success: true,
             data: updatedProduct,
         });
@@ -73,6 +73,13 @@ export const putProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({
+            success: false,
+            message: 'Invalid product ID',
+        });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({
@@ -81,9 +88,9 @@ export const deleteProduct = async (req, res) => {
         });
     } catch (error) {
         console.log('error in deleting product:', error.message);
-        res.status(404).json({
+        res.status(500).json({
             success: false,
-            message: 'Product not found',
+            message: 'Server Error',
         });
     }
 };
